LPS-55052 Since the values are not related to the request but to the discussion taglib I have renamed it to DiscussionTaglibHelper.
