+++ portal-web/docroot/html/taglib/ui/discussion/init.jsp
@@ -21,6 +21,7 @@ page import="com.liferay.portal.kernel.comment.context.CommentTreeDisplayContext
 page import="com.liferay.portlet.messageboards.comment.context.MBCommentSectionDisplayContext" %><%@
 page import="com.liferay.portlet.messageboards.comment.context.MBCommentTreeDisplayContext" %><%@
 page import="com.liferay.portlet.messageboards.comment.context.util.DiscussionRequestHelper" %><%@
+page import="com.liferay.portlet.messageboards.comment.context.util.DiscussionTaglibHelper" %><%@
 page import="com.liferay.portlet.messageboards.model.MBDiscussion" %><%@
 page import="com.liferay.portlet.messageboards.model.MBMessageDisplay" %><%@
 page import="com.liferay.portlet.messageboards.model.MBTreeWalker" %><%@
